fix: several fixes in windows forms app
frm commission fixes
frm courses fixes
frm student-courses fixes
frm my-courses fixes
frm academic-status fixes
frm qualify courses fixes

+++ Academia/ApplicationCore/Model/StudentCourse.cs
@@ -21,7 +21,7 @@ namespace ApplicationCore.Model
 
         [Column("nota")]
         [AllowNull]
-        public int? Grade { get; set; }
+        public decimal? Grade { get; set; }
 
         [Column("id_alumno")]
         public int UserId { get; set; }
@@ -31,5 +31,5 @@ namespace ApplicationCore.Model
         [Column("id_curso")]
         public int CourseId { get; set; }
         public Course? Course { get; set; }
-    }
+	}
 }
